Были добавлены валидаторы, исключения и valueobjects. Изменены Entities
