--- SneakerShop_Service/Domain/Program.cs
@@ -1,10 +1,0 @@
-﻿namespace Domain
-{
-    internal class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello, World!");
-        }
-    }
-}
