+++ SneakerShop_Service/Domain/SneakerShop.Domain/Domain.Enums/Color.cs
@@ -1,0 +1,17 @@
+﻿namespace SneakerShop.Domain.Domain.Enums;
+
+public enum Color
+{
+    Black,
+    White,
+    Red,
+    Blue,
+    Green,
+    Yellow,
+    Pink,
+    Gray,
+    Brown,
+    Beige,
+    Purple,
+    Orange
+}
