+++ SneakerShop_Service/Domain/SneakerShop.Domain/Domain.Enums/Size.cs
@@ -1,0 +1,12 @@
+﻿namespace SneakerShop.Domain.Domain.Enums;
+
+public enum Size
+{
+    XS,
+    S,
+    M,
+    L,
+    XL,
+    XXL,
+    XXXL
+}
