+++ SneakerShop_Service/Domain/SneakerShop.Domain/Domain.Enums/Сurrency.cs
@@ -1,0 +1,8 @@
+﻿namespace SneakerShop.Domain.Domain.Enums;
+
+public enum Currency
+{
+    RUB,
+    USD,
+    EUR
+}
