+++ SneakerShop_Service/Domain/SneakerShop.Domain/SneakerShop.Domain.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\SneakerShop.ValueObject\SneakerShop.ValueObject.csproj" />
+  </ItemGroup>
+
 </Project>
