+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/AdditionalImages.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class AdditionalImages(string? value) : ValueObject<string>(new AdditionalImagesValidator(), value ?? string.Empty);
