+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/BrandName.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class BrandName(string value) : ValueObject<string>(new BrandNameValidator(), value);
