+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/Description.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class Description(string? value) : ValueObject<string>(new DescriptionValidator(), value ?? string.Empty);
