+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/Exceptions/ArgumentNullOrWhiteSpaceException.cs
@@ -1,0 +1,4 @@
+﻿namespace SneakerShop.ValueObject.Exceptions;
+
+public class ArgumentNullOrWhiteSpaceException(string paramName)
+    : ArgumentNullException(paramName, $"The \"{paramName}\" mustn't be null, empty or consists only of white-space characters.");
