+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/Exceptions/ArgumentShortValueException.cs
@@ -1,0 +1,8 @@
+﻿namespace SneakerShop.ValueObject.Exceptions;
+
+public class ArgumentShortValueException(string paramName, string value, int minLength)
+    : FormatException($"The \"{paramName}\" length {value.Length} less than minimum allowed length {minLength}")
+{
+    public string Value => value;
+    public int MinLength => minLength;
+}
