+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/Exceptions/ValidatorNullException.cs
@@ -1,0 +1,4 @@
+﻿namespace SneakerShop.ValueObject.Exceptions;
+
+public class ValidatorNullException(string paramName)
+: ArgumentNullException(paramName, $"Validator \"{paramName}\" must be specified for type.");
