+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/LogoUrl.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class LogoUrl(string value) : ValueObject<string>(new LogoUrlValidator(), value);
