+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/MainImageUrl.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class MainImageUrl(string value) : ValueObject<string>(new MainImageUrlValidator(), value);
