+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/Nickname.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class Nickname(string value) : ValueObject<string>(new NicknameValidator(), value);
