+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/ProductName.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class ProductName(string value) : ValueObject<string>(new ProductNameValidator(), value);
