+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/ProductNameAtCart.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class ProductNameAtCart(string value) : ValueObject<string>(new ProductNameAtCartValidator(), value);
