+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/Sku.cs
@@ -1,0 +1,6 @@
+﻿using SneakerShop.ValueObject.Base;
+using SneakerShop.ValueObject.Validators;
+
+namespace SneakerShop.ValueObject;
+
+public class Sku(string value) : ValueObject<string>(new SkuValidator(), value);
