+++ SneakerShop_Service/Domain/SneakerShop.ValueObject/SneakerShop.ValueObject.csproj
