--- SneakerShop_Service/SneakerShop.Domain/Domain.Enums/Color.cs
@@ -1,18 +1,0 @@
-﻿namespace Domain.Enums
-{
-    public enum Color
-    {
-        Black,
-        White,
-        Red,
-        Blue,
-        Green,
-        Yellow,
-        Pink,
-        Gray,
-        Brown,
-        Beige,
-        Purple,
-        Orange
-    }
-}
