--- SneakerShop_Service/SneakerShop.Domain/Domain.Enums/Size.cs
@@ -1,13 +1,0 @@
-﻿namespace Domain.Enums
-{
-    public enum Size
-    {
-        XS,
-        S,
-        M,
-        L,
-        XL,
-        XXL,
-        XXXL
-    }
-}
