--- SneakerShop_Service/SneakerShop.Domain/Domain.Enums/Сurrency.cs
@@ -1,9 +1,0 @@
-﻿namespace Domain.Enums
-{
-    public enum Currency
-    {
-        RUB,
-        USD,
-        EUR
-    }
-}
