--- SneakerShop_Service/SneakerShop.ValueObject/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace SneakerShop.ValueObject
-{
-    public class Class1
-    {
-
-    }
-}
